Tesseract_v2/requirements.txt
asgiref==3.10.0
channels==4.3.1
channels_redis==4.3.0
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
msgpack==1.1.2
PyJWT==2.10.1
redis==7.0.1
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
tzdata==2025.2
